[SITES-41279] image v3 performance improvment
